refactor: linkpick -> techpick 변경
