--- backend/src/main/java/kernel360/linkpick/LinkpickApplication.java
@@ -1,13 +1,0 @@
-package kernel360.linkpick;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LinkpickApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LinkpickApplication.class, args);
-	}
-
-}
