--- backend/src/main/java/kernel360/linkpick/tag/model/entity/Tag.java
@@ -1,40 +1,0 @@
-package kernel360.linkpick.tag.model.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import kernel360.linkpick.common.model.entity.CreatedAndUpdatedTimeColumn;
-import kernel360.linkpick.user.model.entity.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Table(name = "tag")
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Tag extends CreatedAndUpdatedTimeColumn {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "tag_id")
-	private Long tagId;
-
-	@Column(name = "name", nullable = false)
-	private String name;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	public Tag(String name, User user) {
-		this.name = name;
-		this.user = user;
-	}
-}
