+++ backend/src/main/java/kernel360/techpick/link/model/entity/Link.java
@@ -1,0 +1,33 @@
+package kernel360.techpick.link.model.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Lob;
+import jakarta.persistence.Table;
+import kernel360.techpick.common.model.entity.CreatedAndUpdatedTimeColumn;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Table(name = "link")
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Link extends CreatedAndUpdatedTimeColumn {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "link_id")
+	private Long linkId;
+
+	@Lob
+	@Column(name = "url", nullable = false, columnDefinition = "BLOB")
+	private String url;
+
+	public Link(String url) {
+		this.url = url;
+	}
+}
